add init and install targets to Makefile
The init target is intended to be used to initialize a freshly cloned
repository working directory.

To do this in the right node environment, the riff-docker Makefile has
2 new targets, build-init-image and init-rtc. init-rtc uses the docker
image created by build-init-image which will run `make init` on the bound
volume, in the case of init-rtc that is the riff-rtc working directory.