add arg to specify rtc-build tag in Dockerfile-prod-server
Turns out that only ARGS before the 1st FROM are usable in any of the
FROM directives. Therefore I moved the existing ARG before the comment
for the 1st stage instead of after it in the other Dockerfiles, even
when it was the only ARG used by a FROM directive (for consistency
and clarity).