change dev webpack to not start analyzer server or open report
The prod webpack is currently broken so it is important that the dev build
finish building w/o intervention.

It may be best to leave this even after the prod webpack is working again.
I'd think it would be better to add an npm script to open the bundle analyzer
report (and maybe a Makefile target) so you don't have to shutdown a server/
close a page in the browser every time you do a dev build.