clean up server.js; remove unused pkg, remove functions that will be in lti router, deal w/ misdirected routes
the lti router will handle the lti route including (eventually) figuring out the group(room name)
for the connecting user.

the express-session pkg was reporting a warning because the resave and saveUnitialized options must now
be specified (because the default is eventually going to change)

The routes /api/videodata and /api/signalmaster are sometimes specified in the client config because
that route is expected to be redirected by a reverse proxy before it gets to this server. If it gets
here we'll set the status to 404 as that is friendlier.

For the /api/lti/launch route, the oath_consumer_key will be in the post request body with that name.