revert and lock material-icons-react package to version 1.0.2
THIS SHOULD NOT STAY REVERTED/LOCKED ANY LONGER THAN NECESSARY. Keep up w/ the issue
below and start using the latest version once the issue is fixed.

There were bugs introduced in the material-icons-react 1.0.3 patch version. 1.0.4 fixed
one of them (see https://github.com/logtrace/material-icons-react/issues/28) which solved
some of our missing icon issues, but not all. It seems there are more issues, I got a warning
in our Dashboard (see below) which I've added in a comment to an issue on the
material-icons-react project (see https://github.com/logtrace/material-icons-react/issues/36).

```
Warning: Can't call setState (or forceUpdate) on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in the componentWillUnmount method. in MaterialIcon (created by DashboardView) in a (created by DashboardView) in div (created by DashboardView)
```