Made the POST route handler for lti launches support arbitrary trailing paths; changed the config from /api/lti/config
The intention is that a POST to /lti/launch/:route* will redirect to the SPA's matching route when it loads.
In particular we want to support `/lti/launch/room` and `/lti/launch/riffs`, but the code now supports
any trailing route.

To go along with that, the `/api/lti/config` endpoint has been modified to accept a path query parameter. If
not supplied it will default to `/room`. If you want the config for the riff dashboard you would GET
`/api/lti/config?path=/riffs`. If you want the config for the riff platform home you would use
`/api/lti/config?path=/`.