use webRtcRoom instead of roomName for internal identification of meeting
The webRtcRoom is the name for the video chat meeting (used by both webrtc and the data server). It may differ from
the room name displayed to the user, in particular an LTI authenticated user, because in their case we append
the course Id to their LMS group name to make sure the meeting doesn't get combined w/ students in a different
course with the same group name.